gpio: add sysfs gpio only support
With detection of the gpio cdev interface a part of CMakeLists.txt,
add a define to version.h which for explicitly enabling the gpio cdev
interface. This will allow c-periphery to be compiled for older
platforms which only support the gpio sysfs interface (currently
toolchains with 4.4 headers).

Typically toolchains have older headers than their platform and since
the 4.4 kernel is being maintained until Feb 2022, this means there
will still be new toolchains that are using kernel headers older than
4.8.

Signed-off-by: Ryan Barnett <ryanbarnett3@gmail.com>